begin route splitting
